requirements.txt
# Core Dependencies
boto3>=1.28.0
botocore>=1.31.0

# CLI Interface
click>=8.1.0

# Data Processing
pandas>=2.0.0

# Template Engine
Jinja2>=3.1.0

# HTTP Requests (if needed)
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time Utilities
python-dateutil>=2.8.0

# Optional: Web Framework (for dashboard)
flask>=2.3.0
# OR
# fastapi>=0.100.0
# uvicorn>=0.23.0

# Optional: AI/ML (for advanced features)
# openai>=1.0.0  # If using OpenAI API

